Fix integer precision warnings
- Remove unnecessary cast from unsigned int to unsigned long in dwarf_crc32.c
- Change count from unsigned long to unsigned in dwarf_debuglink.c to match parameter type
- Change dos_sig from unsigned to unsigned long in dwarf_object_detector.c to match return type
- Change new_entry_index from unsigned to unsigned long in dwarf_tsearchhash.c for consistency

These changes fix implicit conversion warnings where unsigned long values were being truncated to unsigned int on 64-bit systems.

+++ src/lib/libdwarf/dwarf_crc32.c
@@ -139,7 +139,7 @@ dwarf_crc32 (Dwarf_Debug dbg,unsigned char *crcbuf,
         /*  Call the public API function so it gets tested too. */
         tcrc = (unsigned int)dwarf_basic_crc32(readbuf,
             (unsigned long)readlenu,
-            (unsigned long)init);
+            init);
         init = tcrc;
         size_left -= readlenu;
     }
